backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "flask>=3.0.3",
    "flask-cors>=4.0.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "flask-admin>=1.6.1",
    "sqlalchemy-utils>=0.41.2",
    "flask-migrate>=4.0.7",
    "python-dotenv>=1.0.1",
    "flask-restx>=1.3.0",
    # https://github.com/astral-sh/rye/issues/745
    "gunicorn>=22.0.0",
    "gevent>=24.2.1",
    "feedparser>=6.0.11",
    "dateparser>=1.2.0",
    "trafilatura>=1.10.0",
    "gne>=0.3.1",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.10"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pytest>=8.2.2",
]
